feat: Update My Account routes to use private paths
- Changed order details link in MyAccountOrderDetails component to point to "/pvt/account/orders".
- Updated ButtonSignIn component to redirect to "/pvt/account" for signed-in users.
- Added new 403 and 404 error pages for unauthorized access in the private account section.
- Created dynamic route handling for unknown paths in the private account section.
- Implemented redirection logic in the My Account index page to navigate to the profile page.
- Developed order details page with server-side data fetching for user orders.
- Created orders list page with pagination and filtering capabilities.
- Added styles for the orders list page.
- Implemented profile page to display user account information.
- Created security settings page for user account management.
- Developed user details page for displaying user roles and organization unit.
- Updated account route constants to reflect new private paths.
