+++ packages/core/src/pages/pvt/account/security.tsx
@@ -30,7 +30,7 @@ import { validateUser } from 'src/sdk/account/validateUser'
 import PageProvider from 'src/sdk/overrides/PageProvider'
 
 import { SecuritySection } from 'src/components/account/security'
-import storeConfig from '../../../discovery.config'
+import storeConfig from '../../../../discovery.config'
 
 /* A list of components that can be used in the CMS. */
 const COMPONENTS: Record<string, ComponentType<any>> = {
@@ -131,7 +131,7 @@ export const getServerSideProps: GetServerSideProps<
   if (security.errors) {
     const statusCode: number = (security.errors[0] as any)?.extensions?.status
     const destination: string =
-      statusCode === 403 ? '/account/403' : '/account/404'
+      statusCode === 403 ? '/pvt/account/403' : '/pvt/account/404'
 
     return {
       redirect: {
